requirements.txt
numpy
pandas==2.2.3
Pillow>=11.0.0
tomli==2.2.1
tomli_w==1.2.0
torch
torchmetrics==1.7.1
torchvision
tqdm==4.67.1
lpips
scikit-image
scikit-learn
PyWavelets>=1.4.1
tensorboard